requirements.txt
streamlit
langgraph
langchain-google-genai
ddgs
python-dotenv
